+++ restli-server/src/main/java/com/linkedin/restli/internal/server/response/ResponseUtils.java
@@ -252,6 +252,10 @@ public class ResponseUtils
     return new RestException(restResponse, cause==null ? null : cause.toString(), cause, writableStackTrace);
   }
 
+  /**
+   * @Deprecated: Use buildStreamException(RestLiResponseException, ContentType) method instead
+   */
+  @Deprecated
   public static StreamException buildStreamException(RestLiResponseException restLiResponseException, StreamDataCodec codec)
   {
     RestLiResponse restLiResponse = restLiResponseException.getRestLiResponse();
@@ -262,6 +266,21 @@ public class ResponseUtils
             : restLiResponse.getStatus().getCode());
 
     EntityStream<ByteString> entityStream = codec.encodeMap(restLiResponse.getDataMap());
+    StreamResponse response = responseBuilder.build(EntityStreamAdapters.fromGenericEntityStream(entityStream));
+    return new StreamException(response, restLiResponseException.getCause());
+  }
+
+  public static StreamException buildStreamException(RestLiResponseException restLiResponseException, ContentType contentType)
+  {
+    RestLiResponse restLiResponse = restLiResponseException.getRestLiResponse();
+    StreamResponseBuilder responseBuilder = new StreamResponseBuilder()
+        .setHeaders(restLiResponse.getHeaders())
+        .setHeader(RestConstants.HEADER_CONTENT_TYPE, contentType.getHeaderKey())
+        .setCookies(CookieUtil.encodeSetCookies(restLiResponse.getCookies()))
+        .setStatus(restLiResponse.getStatus() == null ? HttpStatus.S_500_INTERNAL_SERVER_ERROR.getCode()
+            : restLiResponse.getStatus().getCode());
+
+    EntityStream<ByteString> entityStream = contentType.getStreamCodec().encodeMap(restLiResponse.getDataMap());
     StreamResponse response = responseBuilder.build(EntityStreamAdapters.fromGenericEntityStream(entityStream));
     return new StreamException(response, restLiResponseException.getCause());
   }